refactor(video): share one owner-video action sheet between the grids
`ProfileVideosGrid` and `ComposableVideoGrid` presented the same edit/delete
actions through two independent implementations. The profile one used
`VineBottomSheet` with `VineTheme` typography; the composable one hand-rolled a
raw `showModalBottomSheet` with eight inline `TextStyle`s, its own header row
and close button, and a separate delete path that reported through raw
`SnackBar`s in `VineTheme.warning` / `VineTheme.error` with a hardcoded two
second "Deleting content..." toast. The same delete told the user different
things depending on which grid they started it from.

Extract `showOwnerVideoActionsSheet` into `lib/widgets/owner_video_actions_sheet.dart`,
alongside the sheet body and action tile promoted unchanged out of the profile
grid. The helper owns presentation and the whole delete orchestration —
confirm, publish, cleanup completion, snackbar — so both surfaces now report
identically through `DivineSnackbarContainer` with the shared localized copy.

The surfaces differ in one real way, so the helper takes it as a parameter
rather than assuming it: after a successful delete the profile grid refreshes
its feed to drop the tile without waiting for relay propagation, while the
mixed-owner grids have no feed cubit to refresh. That is `onDeleted`.

This follows the three previous consolidations of this feature — the delete
confirmation dialog, the cleanup feedback, and `OwnerVideoActionsCubit` — all
of which are functions or classes every surface calls, so no new architectural
shape is introduced. `feed_settings_menu.dart` is deliberately untouched: it is
an overlay popover, a different UX pattern, and #6327 names it as the reference
rather than a target.

Net effect on the design-system ratchets, all in `composable_video_grid.dart`:
raw dialogs 1 -> 0, raw `TextStyle(` 8 -> 0, Material buttons 1 -> 0. Baselines
are regenerated in the next commit.

Refs #6327
