+++ mobile/lib/widgets/composable_video_grid.dart
@@ -6,23 +6,20 @@ import 'dart:async';
 import 'package:divine_ui/divine_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart' show ScrollCacheExtent;
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
+import 'package:go_router/go_router.dart';
 import 'package:models/models.dart' hide AspectRatio;
 import 'package:openvine/blocs/owner_video_actions/owner_video_actions_cubit.dart';
-import 'package:openvine/extensions/modal_pop_extension.dart';
 import 'package:openvine/l10n/l10n.dart';
 import 'package:openvine/mixins/scroll_pagination_mixin.dart';
 import 'package:openvine/providers/app_providers.dart';
 import 'package:openvine/providers/creator_delete_enforcement_providers.dart';
 import 'package:openvine/providers/nostr_client_provider.dart';
-import 'package:openvine/utils/delete_result_localization.dart';
-import 'package:openvine/utils/owner_video_cleanup_feedback.dart';
+import 'package:openvine/screens/video_metadata/video_metadata_edit_screen.dart';
 import 'package:openvine/widgets/branded_loading_indicator.dart';
 import 'package:openvine/widgets/feed_refresh_control.dart';
-import 'package:openvine/widgets/owner_video_delete_confirmation_dialog.dart';
-import 'package:openvine/widgets/show_edit_dialog_for_video.dart';
+import 'package:openvine/widgets/owner_video_actions_sheet.dart';
 import 'package:openvine/widgets/user_name.dart';
 import 'package:openvine/widgets/video_thumbnail_widget.dart';
 
@@ -290,247 +287,16 @@ class _ComposableVideoGridState extends ConsumerState<ComposableVideoGrid>
   ///
   /// Ownership is decided in [_buildGrid] so a non-owned tile never wires the
   /// gesture — and therefore never advertises the action to assistive tech.
-  void _showVideoContextMenu(BuildContext context, VideoEvent video) {
-    showModalBottomSheet<void>(
-      context: context,
-      backgroundColor: context.vineColors.background,
-      builder: (sheetContext) => BlocProvider.value(
-        value: _ownerVideoActionsCubit,
-        child: BlocBuilder<OwnerVideoActionsCubit, OwnerVideoActionsState>(
-          builder: (_, state) {
-            final operation = state.forVideo(video.id);
-            final isDeletePublishing =
-                operation.deleteStatus == OwnerVideoDeleteStatus.deleting;
-            final isDeleteInProgress =
-                isDeletePublishing ||
-                operation.cleanupStatus == OwnerVideoCleanupStatus.inProgress;
-            return SafeArea(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  // Header
-                  Padding(
-                    padding: const EdgeInsets.all(16),
-                    child: Row(
-                      children: [
-                        DivineIcon(
-                          icon: DivineIconName.dotsThreeVertical,
-                          color: sheetContext.vineColors.primaryText,
-                        ),
-                        const SizedBox(width: 12),
-                        Expanded(
-                          child: Text(
-                            sheetContext.l10n.videoGridOptionsTitle,
-                            style: TextStyle(
-                              color: sheetContext.vineColors.primaryText,
-                              fontSize: 18,
-                              fontWeight: FontWeight.bold,
-                            ),
-                          ),
-                        ),
-                        IconButton(
-                          onPressed: () => sheetContext.popModalIfMounted(),
-                          icon: DivineIcon(
-                            icon: DivineIconName.x,
-                            color: sheetContext.vineColors.secondaryText,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-
-                  // Edit option
-                  ListTile(
-                    leading: Container(
-                      width: 40,
-                      height: 40,
-                      decoration: BoxDecoration(
-                        color: sheetContext.vineColors.card,
-                        borderRadius: BorderRadius.circular(8),
-                      ),
-                      child: DivineIcon(
-                        icon: DivineIconName.pencilSimple,
-                        color: sheetContext.vineColors.accentPositive,
-                        size: 20,
-                      ),
-                    ),
-                    title: Text(
-                      sheetContext.l10n.videoGridEditVideo,
-                      style: TextStyle(
-                        color: sheetContext.vineColors.primaryText,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                    subtitle: Text(
-                      sheetContext.l10n.videoGridEditVideoSubtitle,
-                      style: TextStyle(
-                        color: sheetContext.vineColors.secondaryText,
-                        fontSize: 12,
-                      ),
-                    ),
-                    enabled: !isDeleteInProgress,
-                    onTap: isDeleteInProgress
-                        ? null
-                        : () {
-                            if (_ownerVideoActionsCubit.isDeleteInProgress(
-                              video.id,
-                            )) {
-                              return;
-                            }
-                            if (!sheetContext.popModalIfMounted()) return;
-                            showEditDialogForVideo(context, video);
-                          },
-                  ),
-
-                  // Delete option
-                  ListTile(
-                    leading: Container(
-                      width: 40,
-                      height: 40,
-                      decoration: BoxDecoration(
-                        color: sheetContext.vineColors.card,
-                        borderRadius: BorderRadius.circular(8),
-                      ),
-                      child: isDeleteInProgress
-                          ? const CircularProgressIndicator(strokeWidth: 2)
-                          : const DivineIcon(
-                              icon: DivineIconName.trash,
-                              color: VineTheme.error,
-                              size: 20,
-                            ),
-                    ),
-                    title: Text(
-                      sheetContext.l10n.videoGridDeleteVideo,
-                      style: TextStyle(
-                        color: sheetContext.vineColors.primaryText,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                    subtitle: Text(
-                      sheetContext.l10n.videoGridDeleteVideoSubtitle,
-                      style: TextStyle(
-                        color: sheetContext.vineColors.secondaryText,
-                        fontSize: 12,
-                      ),
-                    ),
-                    enabled: !isDeleteInProgress,
-                    onTap: isDeleteInProgress
-                        ? null
-                        : () {
-                            if (!sheetContext.popModalIfMounted()) return;
-                            _showDeleteConfirmation(context, video);
-                          },
-                  ),
-
-                  const SizedBox(height: 16),
-                ],
-              ),
-            );
-          },
+  Future<void> _showVideoContextMenu(BuildContext context, VideoEvent video) =>
+      showOwnerVideoActionsSheet(
+        context: context,
+        video: video,
+        cubit: _ownerVideoActionsCubit,
+        onEditRequested: () => context.push(
+          VideoMetadataEditScreen.pathFor(video.id),
+          extra: video,
         ),
-      ),
-    );
-  }
-
-  /// Show delete confirmation dialog
-  Future<void> _showDeleteConfirmation(
-    BuildContext context,
-    VideoEvent video,
-  ) async {
-    final confirmed = await showOwnerVideoDeleteConfirmationDialog(context);
-
-    if (confirmed && context.mounted) {
-      await _deleteVideo(context, video);
-    }
-  }
-
-  /// Delete the video and confirm Divine-controlled media cleanup.
-  Future<void> _deleteVideo(BuildContext context, VideoEvent video) async {
-    try {
-      if (_ownerVideoActionsCubit.isDeleteInProgress(video.id)) return;
-      // Show loading snackbar
-      if (context.mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Row(
-              children: [
-                SizedBox(
-                  width: 16,
-                  height: 16,
-                  child: CircularProgressIndicator(
-                    strokeWidth: 2,
-                    color: context.vineColors.primaryText,
-                  ),
-                ),
-                const SizedBox(width: 12),
-                Text(context.l10n.videoGridDeletingContent),
-              ],
-            ),
-            backgroundColor: VineTheme.warning,
-            duration: const Duration(seconds: 2),
-          ),
-        );
-      }
-
-      final start = await _ownerVideoActionsCubit.deleteVideo(video);
-      if (start == OwnerVideoDeleteStart.busy) return;
-      if (!context.mounted) return;
-      final operation = _ownerVideoActionsCubit.state.forVideo(video.id);
-      if (operation.deleteStatus == OwnerVideoDeleteStatus.success) {
-        showOwnerVideoCleanupCompletion(
-          context,
-          _ownerVideoActionsCubit,
-          video.id,
-        );
-      }
-
-      if (context.mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Row(
-              children: [
-                Icon(
-                  operation.deleteStatus == OwnerVideoDeleteStatus.success
-                      ? Icons.check_circle
-                      : Icons.error,
-                  color: context.vineColors.primaryText,
-                ),
-                const SizedBox(width: 12),
-                Expanded(
-                  child: Text(
-                    operation.deleteStatus == OwnerVideoDeleteStatus.success
-                        ? localizedOwnerVideoDeleteSuccessMessage(
-                            context,
-                            operation,
-                          )
-                        : operation.deleteResult == null
-                        ? context.l10n.shareMenuDeleteFailedGeneric
-                        : localizedDeleteFailureMessage(
-                            context,
-                            operation.deleteResult!,
-                          ),
-                  ),
-                ),
-              ],
-            ),
-            backgroundColor:
-                operation.deleteStatus == OwnerVideoDeleteStatus.success
-                ? VineTheme.vineGreen
-                : VineTheme.error,
-          ),
-        );
-      }
-    } catch (e) {
-      if (context.mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text(context.l10n.shareMenuDeleteFailedGeneric),
-            backgroundColor: VineTheme.error,
-          ),
-        );
-      }
-    }
-  }
+      );
 }
 
 class _VideoItem extends StatelessWidget {
